chore(attach-mock): rename mocks folder to test
