+++ src/core/data/middlewares/attach-mock/test/test-provider-mocks.ts
@@ -1,0 +1,9 @@
+import { responseData } from 'core/data/middlewares/attach-mock/test/response-data';
+
+export default {
+	GET: [
+		{
+      response: responseData
+		}
+	]
+};
